# Improve camera preview performance