--- a/src-tauri/Cargo.toml
+++ b/src-tauri/Cargo.toml
@@ -47,6 +47,7 @@ tauri-plugin-os = "2"
 tauri-plugin-dialog = "2"
 tauri-plugin-log = "2"
 log = "0.4"
+rayon = "1.10.0"
 
 [dependencies.nokhwa]
 git = "https://github.com/l1npengtul/nokhwa.git"